+++ packages/typedoc-plugin-markdown/src/router/index.ts
@@ -4,5 +4,5 @@
  * @module
  */
 export * from './markdown-router.js';
-export * from './routers/kind-structure-router.js';
+export * from './routers/member-router.js';
 export * from './routers/module-router.js';